+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/css/index.css';
 import reportWebVitals from './reportWebVitals';
 import App from './App';
