+++ src/layout/sections/herosection/logo/Logo.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import logo from "../../components/assets/images/logo.jpg";
+import logo from "../../../../assets/images/logo.jpg";
 import { myTheme } from "../../../../styles/global/MyTheme.styled";
 // import { Icon } from "../icon/Icon";
